+++ CHANGELOG.md
@@ -6,7 +6,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
-## [Unreleased](https://github.com/inspirum/balikobot-php/compare/v7.5.1...master)
+## [Unreleased](https://github.com/inspirum/balikobot-php/compare/v7.6.0...master)
+
+
+## [v7.6.0 (2023-05-29)](https://github.com/inspirum/balikobot-php/compare/v7.5.1...v7.6.0)
+### Changed
+- Adjusted `guzzlehttp/psr7` version constraint to `^1.5 || ^2.0`
+- Adjusted `psr/http-message` version constraint to `^1.1 || ^2.0`
 
 
 ## [v7.5.1 (2023-05-15)](https://github.com/inspirum/balikobot-php/compare/v7.5.0...v7.5.1)